Attempted WIP Link Switch
Wrote a function to switch the link for mangas labelled as Work in Progress to a direct link to the latest chapter read. Due to multiple nuances, it did not universally work. I commented the function out and wrote a simple if statement instead that replaces the link to the latest chapter released with the link to the source url.