Manga List Storage Update
The program now saves the list of mangas in a text file called list.txt, rather than a Python list. This change parallels the previous creation of latest.txt and is part of a greater effort to make running and customizing the file more accessible to non-coders.

In the case of an empty saved/list.txt, the program uses the old Python list.

Created the primer() function which helps first time users choose which of the developer's mangas they'd like to keep on their list and change the latest chapter as well as manga status to fit their needs.