Reordering and Update Latest
Reordered the the functions and code to better reflect the actual control flow throughout the document with the top being the start and the bottom as the end.

Fixed an error caused by the dynamic handling with update latest by treating all dynamic mangas like wip or yts. This means their lines in latest.txt must be changed through change_curent() instead.