Unexpected Behavior Fixes
Prevented script from aborting when dealing with an improperly loaded dynamic manga by using a try-except statement at the source.

Wrote another file write at the end of finisher for s() such that dynamic mangas are included in the text as well.

Mangas labelled as yts would show the latest chapter's link. Now this does not happen for wip and yts labelled manga.