Eliminating Clutter
Prior installation led to the creation of an Extracts and Zips folder with MQuicker found within Extracts and the MQ.zip used to create it found within. Now, one MQuicker folder is created in the C:\ drive.

The Desktop shortcut for MQuicker is now automatically moved to the Desktop.

Created a setup for Mac users that already have python configured.