README, SetUp, and Color
Updated the README for clarity and correctness.

Changed the setup ps1 files to make all Write-Host outputs green and changed the path to Desktop to $env:USERPROFILE + "\Desktop" with the understanding that most people do not use \OneDrive

The typo LIGHTGREEN_Ex (instead of EX) broke primer() so I corrected that.