Input Verification & Error Handling
Created various checks for invalid input. One such case is the verify_status() function and another is the assertion that a link must start with https:// and have a period as well as the specified source code needing to have suppoprt already for added manga.

Built a set_changes() function to update the live mangas and current lists after a call of primer(), add(), or change_current().

Added a try-except statement within finder() to output a latest chapter released of "-1" if the website is down during the function call.